requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
httpx==0.28.1
feedparser==6.0.12
python-dotenv==1.2.1
beautifulsoup4==4.14.3
trafilatura==2.0.0
yfinance==1.2.0
python-dateutil==2.9.0.post0
requests==2.32.5
